refactor(boot): extract AuthGate from main.jsx into its own module
WIP checkpoint — the extraction half of #285. Tests and the coverage
threshold entry follow in the next commit; do not merge before them.

Moves C, Login, SignOutButton, InstallButton and AuthGate out of
main.jsx into web/src/AuthGate.jsx verbatim — a byte-identical move
across all 283 lines, the sole delta being the `export default`
keyword on AuthGate. main.jsx drops to 64 lines of pure bootstrap:
initSentry, the theme <style> injection, the Capacitor branch, the
queryClient construction and the root render.

All four components travel together because AuthGate renders each of
them and they share the private C token map. InstallButton in
particular cannot stay behind: it is one of the three children of the
flat fragment whose child index the splash animation depends on.

Refs #285
